feat(carousel): add bullet base and styled components, general organization
