+++ jsconfig.json
@@ -7,7 +7,8 @@
             "#components": ["./src/components"],
             "#containers": ["./src/containers"],
             "#context": ["./src/context"],
-            "#utils": ["./src/utils"]
+            "#utils": ["./src/utils"],
+            "#base": ["./src/base"]
         }
     },
     "include": ["./src/**/*"],
