+++ src/base/index.js
@@ -1,3 +1,5 @@
+export { default as BaseBullets } from './BaseBullets';
 export { default as BaseCarousel } from './BaseCarousel';
 export { default as BaseScrollSlider } from './BaseScrollSlider';
 export { default as BaseSwipeSlider } from './BaseSwipeSlider';
+export { default as BaseUnevenItemsScrollSlider } from './BaseUnevenItemsScrollSlider';
