+++ src/components/BulletsInfiniteWrapper.js
@@ -1,0 +1,13 @@
+import styled from '@emotion/styled';
+
+const BulletsInfiniteWrapper = styled.div`
+    position: absolute;
+    top: 0;
+    left: 0;
+    transition: transform 200ms;
+    display: flex;
+`;
+
+BulletsInfiniteWrapper.displayName = 'BulletsInfiniteWrapper';
+
+export default BulletsInfiniteWrapper;
