+++ stories/SnapSliderCarousel.stories.jsx
@@ -1,11 +1,42 @@
 import './Carousel.css';
-import { Arrow, Carousel, SnapSlider } from '../src';
+import { Arrow, Bullets, Carousel, SnapSlider } from '../src';
 import React from 'react';
 
 export default {
     component: Carousel,
     title: 'Carousel/SnapSlider',
 };
+
+const children = [
+    <a key={1} className="item first">
+        <svg viewBox="0 0 7 4"></svg>
+        <p>1</p>
+    </a>,
+    <a key={2} className="item second">
+        <svg viewBox="0 0 7 4"></svg>
+        <p>2</p>
+    </a>,
+    <a key={3} className="item third">
+        <svg viewBox="0 0 7 4"></svg>
+        <p>3</p>
+    </a>,
+    <a key={4} className="item fourth">
+        <svg viewBox="0 0 7 4"></svg>
+        <p>4</p>
+    </a>,
+    <a key={5} className="item fifth">
+        <svg viewBox="0 0 7 4"></svg>
+        <p>5</p>
+    </a>,
+    <a key={6} className="item sixth">
+        <svg viewBox="0 0 7 4"></svg>
+        <p>6</p>
+    </a>,
+    <a key={7} className="item seventh">
+        <svg viewBox="0 0 7 2"></svg>
+        <p>7</p>
+    </a>,
+];
 
 const Template = (args) => (
     <div className="area">
@@ -19,57 +50,9 @@ const Template = (args) => (
                 animationDuration={args.animationDuration}
                 animationTimingFunction={args.animationTimingFunction}
             >
-                <a
-                    href="#"
-                    className="card"
-                    style={{
-                        backgroundColor: '#72147E',
-                    }}
-                >
-                    <svg viewBox="0 0 7 2"></svg>
-                    <p>1</p>
-                </a>
-                <a
-                    href="#"
-                    className="card"
-                    style={{
-                        backgroundColor: '#F21170',
-                    }}
-                >
-                    <svg viewBox="0 0 7 2"></svg>
-                    <p>2</p>
-                </a>
-                <a
-                    href="#"
-                    className="card"
-                    style={{
-                        backgroundColor: '#FA9905',
-                    }}
-                >
-                    <svg viewBox="0 0 7 2"></svg>
-                    <p>3</p>
-                </a>
-                <a
-                    href="#"
-                    className="card"
-                    style={{
-                        backgroundColor: '#FF5200',
-                    }}
-                >
-                    <svg viewBox="0 0 7 2"></svg>
-                    <p>4</p>
-                </a>
-                <a
-                    href="#"
-                    className="card"
-                    style={{
-                        backgroundColor: '#CF0000',
-                    }}
-                >
-                    <svg viewBox="0 0 7 2"></svg>
-                    <p>5</p>
-                </a>
+                {children}
             </SnapSlider>
+            <Bullets />
             <Arrow flow={'prev'}>
                 {({ onClick }) => <button onClick={onClick}>Previous</button>}
             </Arrow>
@@ -81,8 +64,8 @@ const Template = (args) => (
     </div>
 );
 
-export const SnapSliderCarousel = Template.bind({});
-SnapSliderCarousel.args = {
+export const SnapCarousel = Template.bind({});
+SnapCarousel.args = {
     isInfinite: false,
     isRTL: false,
     itemsToShow: 1.25,
